backend/pyproject.toml
[tool.poetry]
name = "jroots-backend"
version = "0.2.0"
description = "Backend for Jewish OCR Search"
authors = ["Michael Akushsky <michael.akushsky@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
pydantic = { extras = ["email"], version = "^2.5.3" }
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
Pillow = "^10.2.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
httpx = "^0.25.0"
tenacity = "^8.2.2"
certifi = "^2024.9.14"
python-logging-loki = "^0.3.1"
python-json-logger = "^3.3.0"
colorlog = "^6.7.0"
alembic = "^1.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.0"
ruff = "^0.0.292"
mypy = "^1.5.0"
pre-commit = "^3.0.0"
black = "^24.2.0"
respx = "^0.22.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
